+++ src/WeeklyAlarm.h
@@ -32,38 +32,33 @@ Permission is hereby granted, free of charge, to any person obtaining a copy of
 
 #ifdef __NEWLIB__
     #include <time.h>
-    #define WA_NOW()           time(nullptr)
-    #define WA_BREAKTIME(t, e) do { time_t _t = (t); struct tm* _tmp = localtime(&_t); (e) = *_tmp; } while(0)
-    #define WA_MAKETIME(e)     mktime(&e)
-    typedef struct tm          WA_TimeElements;
-    #define WA_WDAY            tm_wday
-    #define WA_HOUR            tm_hour
-    #define WA_MINUTE          tm_min
-    #define WA_DAY             tm_mday
-    #define WA_MONTH           tm_mon
-    #define WA_YEAR            tm_year
-    #define WA_WDAY_OFFSET 0  // 0=Sunday
+    inline time_t now() {return time(nullptr);}
+    typedef struct
+    {
+        uint16_t Year;    // year since 1970
+        uint8_t Month;    // 1-12
+        uint8_t Day;      // 1-31
+        uint8_t Wday;     // Wday: Sunday=1 ... Saturday=7
+        uint8_t Hour;
+        uint8_t Minute;
+        uint8_t Second;
+    }tmElements_t, TimeElements, *tmElementsPtr_t;
+
+    void breakTime(time_t t, tmElements_t &tm);
+
+    time_t makeTime(const tmElements_t &e);
     typedef enum {
       dowInvalid, dowSunday, dowMonday, dowTuesday, dowWednesday, dowThursday, dowFriday, dowSaturday
     } timeDayOfWeek_t;
 #else
-    //#include "TimeLib.h"
-    //include Time this way if you using PlatformIo
+#ifdef WEEKLYALARM_USE_LOCAL_TIMELIB
+//if using PlatformIO, the include path is different,
+//Add in platform.ini: build_flags = -DWEEKLYALARM_USE_LOCAL_TIMELIB
     #include "../Time/TimeLib.h"
-    #include "../Time/TimeLib.h"
-    #define WA_NOW()           now()
-    #define WA_BREAKTIME(t, e) breakTime(t, e)
-    #define WA_MAKETIME(e)     makeTime(e)
-    typedef TimeElements       WA_TimeElements;
-    #define WA_WDAY            Wday
-    #define WA_HOUR            Hour
-    #define WA_MINUTE          Minute
-    #define WA_DAY             Day
-    #define WA_MONTH           Month
-    #define WA_YEAR            Year
-    #define WA_WDAY_OFFSET 1  // 1=Sunday (dowSunday)
-#endif
-
+#else
+    #include <TimeLib.h>
+#endif //WEEKLYALARM_USE_LOCAL_TIMELIB
+#endif //__NEWLIB__
 
 
 #define ALARM_ENABLE_MASK (timeDayOfWeek_t)0
@@ -79,8 +74,8 @@ class AlarmNode {
   friend class WeeklyAlarm;
   bool reset();
   bool isEnable();
-  int8_t getDayToGo(WA_TimeElements &now);
-  bool todaysTimeIsPast(WA_TimeElements &now, WA_TimeElements &alrm) const; 
+  int8_t getDayToGo(tmElements_t &now);
+  bool todaysTimeIsPast(tmElements_t &now, tmElements_t &alrm) const; 
   uint8_t dayEnable;
   virtual void callback() = 0; 
   AlarmNode *nextAlarm;
@@ -269,16 +264,37 @@ public:
     return alarm->dayEnable;
   }
 
+/**
+ * @brief Schedule a single alarm at a specific epoch time.
+ *
+ * The alarm will execute its callback once when the target time is reached.
+ * The callback is responsible for scheduling a new alarm if a recurring
+ * behavior is required.
+ * Must be followed by arm() to insert into the active queue.
+ * @param target Absolute target time as a Unix epoch timestamp.
+ *
+ * @note This does not calculate the next occurrence. The caller is
+ *       responsible for determining the correct target time.
+ */
+  void setSingleAlarm(time_t target);
 
   /**
    * @brief Get the target time of the alarm
    * @return WA_TimeElements; arduino timeLib: TimeElements, POSIX lib: struct tm
    */
-  WA_TimeElements getTime() const {
-    WA_TimeElements time;
-    WA_BREAKTIME(alarm->target, time);
+  tmElements_t getTime() const {
+    tmElements_t time;
+    breakTime(alarm->target, time);
     return time;
   }
+
+  /**
+  * @brief Insert alarm into sorted queue at current target time.
+  * Use after setSingleAlarm() to complete the arming sequence,
+  * or to re-queue after target has been modified externally.
+  * For normal use, prefer alarmOn() which computes target automatically.
+  */
+  void arm();
 
   /**
    * @brief Build a JSON string of the alarm settings
@@ -297,11 +313,6 @@ private:
   * Pop alarms instance from the list.
   */
   static void pop(AlarmNode *node);
-};
-
-struct AlarmProperties
-{
-  uint8_t daysEnables = 0;
 };
 
 #endif